requirements.txt
confluent-kafka==2.11.0
pyarrow==21.0.0
python-dotenv==1.1.1
PyYAML==6.0.2
